Structure the veymont rewrites into verification and generation packages. Move the permission generation rewrite in the general rewrite package
